large refactor, just to make mypy not gripe
